fix: resolve 7 pre-existing --all-features compile errors (Issue 004 filed for remaining runtime.h blocker)
- feedback.rs: format Vec<u8> via String::from_utf8_lossy + annotate
  mpsc channel type (Display trait missing for bytes)
- proof_cert/wasm_proof_witness.rs: drop P5Derived<'a> lifetime, use
  owned [String; 2] — borrowed strs broke serde Deserialize bound
- percepta/evaluator.rs: hoist dim_order.len() before struct move;
  convert attention_insert_and_query to associated fn taking explicit
  disjoint field refs (&mut attention_entries, position,
  &mut scratch_attn_total) — resolves double &mut self borrow
- skill_opt/buffer.rs: annotate TryFromSliceError on try_into map_err
- pruners/feedback_bandit.rs: fix HashMap key type (usize,usize,usize)
  to (usize,u8,u8) matching ConfiguratorContext field types
- pruners/skill_catalog.rs: papaya has no get_mut on HashMapRef — use
  functional update() with clone (SkillDescriptor is Clone)

Validation: --all-features 10 errors -> 1 (missing runtime.h, Issue 004).
Default features: 3842/3842 tests pass. Fixed-module tests: 12/12 pass.

+++ src/pruners/feedback_bandit.rs
@@ -212,7 +212,7 @@ pub struct FeedbackBandit {
     /// Inner SR²AM configurator bandit (always 4 arms).
     inner: ConfiguratorBandit,
     /// Per-context stats for feedback arms (arms 4-5).
-    feedback_stats: HashMap<(usize, usize, usize), FeedbackContextStats>,
+    feedback_stats: HashMap<(usize, u8, u8), FeedbackContextStats>,
     /// FeedbackBandit configuration.
     config: FeedbackBanditConfig,
     /// Trajectory summary for stall detection.
